make a separate component for the simple timer
